requirements.txt
mysql-connector-python
pandas>=2.2.0
numpy>=1.26.0
scikit-learn>=1.4.0